move laika.ast.MessageFilter to laika.config
