+++ core/shared/src/main/scala/laika/api/builder/ParserBuilder.scala
@@ -18,8 +18,7 @@ package laika.api.builder
 
 import laika.api.MarkupParser
 import laika.api.format.MarkupFormat
-import laika.ast.MessageFilter
-import laika.config.MessageFilters
+import laika.config.{ MessageFilter, MessageFilters }
 
 /** Builder API for Parser instances.
   *
