+++ core/shared/src/test/scala/laika/render/HTMLRendererSpec.scala
@@ -26,10 +26,9 @@ import laika.api.bundle.{
 }
 import laika.ast.Path.Root
 import laika.ast.*
-import laika.ast.MessageFilter
 import laika.ast.sample.{ ParagraphCompanionShortcuts, TestSourceBuilders }
 import laika.ast.styles.StyleDeclarationSet
-import laika.config.{ MessageFilters, TargetFormats, Version, Versions }
+import laika.config.{ MessageFilter, MessageFilters, TargetFormats, Version, Versions }
 import laika.format.HTML
 import laika.parse.GeneratedSource
 import laika.parse.code.CodeCategory
